Update pfatt.sh
These are the small changes I had to make based on various forum posts as well as the issues tracker here. They work across reboots with the existing instructions on a bare metal installation of pfsense. 

As far as I can tell, they simply clear out incomplete configs if they are present, specify a mac address, and change the wpa_supplicant interface to be the physical hardware interface, rather than the netgraph one.